ScoreManagerをタイマー機能のみに変更
